🗃 Remove `visible` when acking jobs
This is a non-breaking change that forms part of the migration path to
improving our performance in https://github.com/reedsy/mongodb-queue/pull/15

When acking a job, we now unset the `visible` property, which makes
`visible` and `deleted` mutually exclusive properties, so that the
presence of one implies the absence of the other.

We can do this, because the only time we query on `visible` is when we
*also* query for `deleted: {$exists: false}`. Similarly, the only times
we query for `deleted: {$exists: true}` are times when we don't query
`visible`.